added generic reward system that appears after vanquishing all of the enemies in a room.  selecting a reward opens the doors.

+++ Assets/Blueprints/PlayerValues.cs
@@ -10,5 +10,5 @@ public class PlayerValues : DamageableConfig
     public float currentHealth = 100;
 
     [Min(0)]
-    public float goldCount = 0;
+    public int goldCount = 0;
 }
